added exception

+++ src/main/java/com/demo/application/product/controller/ProductServiceController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.demo.application.product.exception.ProductNotfoundException;
 import com.demo.application.product.model.Product;
 
 @RestController
@@ -30,28 +31,37 @@ public class ProductServiceController {
       productRepo.put(almond.getId(), almond);
    }
    
-   @RequestMapping(value = "/products/{id}", method = RequestMethod.DELETE)
-   public ResponseEntity<Object> delete(@PathVariable("id") String id) { 
-      productRepo.remove(id);
-      return new ResponseEntity<>("Product is deleted successsfully", HttpStatus.OK);
+   @RequestMapping(value = "/CreateProducts", method = RequestMethod.POST)
+   public ResponseEntity<Object> createProduct(@RequestBody Product product) {
+      productRepo.put(product.getId(), product);
+      return new ResponseEntity<>("Products are created successfully", HttpStatus.CREATED);
    }
    
-   @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
+   @RequestMapping(value = "/GetAllProducts")
+   public ResponseEntity<Object> getProduct() {
+      return new ResponseEntity<>(productRepo.values(), HttpStatus.OK);
+   }
+   @RequestMapping(value = "/GetProduct/{id}")
+   public ResponseEntity<Object> getProductByID(@PathVariable("id") String id) {
+	   if(!productRepo.containsKey(id))throw new ProductNotfoundException();
+      return new ResponseEntity<>(productRepo.get(id).getName(), HttpStatus.OK);
+   }
+   
+   @RequestMapping(value = "/UpdateProduct/{id}", method = RequestMethod.PUT)
    public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product) { 
+	   if(!productRepo.containsKey(id))throw new ProductNotfoundException();
       productRepo.remove(id);
       product.setId(id);
       productRepo.put(id, product);
       return new ResponseEntity<>("Product is updated successsfully", HttpStatus.OK);
    }
    
-   @RequestMapping(value = "/products", method = RequestMethod.POST)
-   public ResponseEntity<Object> createProduct(@RequestBody Product product) {
-      productRepo.put(product.getId(), product);
-      return new ResponseEntity<>("Product is created successfully", HttpStatus.CREATED);
+   @RequestMapping(value = "/DeleteProduct/{id}", method = RequestMethod.DELETE)
+   public ResponseEntity<Object> delete(@PathVariable("id") String id) { 
+	   if(!productRepo.containsKey(id))throw new ProductNotfoundException();
+      productRepo.remove(id);
+      return new ResponseEntity<>("Product is deleted successsfully", HttpStatus.OK);
    }
    
-   @RequestMapping(value = "/products")
-   public ResponseEntity<Object> getProduct() {
-      return new ResponseEntity<>(productRepo.values(), HttpStatus.OK);
-   }
+
 }
